CE-1431 Separate facet for /editor_preview/ and other sanity fixes
