+++ front/scripts/main/controllers/ApplicationController.ts
@@ -16,7 +16,8 @@ App.ApplicationController = Em.Controller.extend({
 			domain: Em.get(Mercury, 'wiki.dbName') || window.location.href.match(/^https?:\/\/(.*?)\./)[1],
 			language: Em.get(Mercury, 'wiki.language'),
 			mainPageTitle: Em.get(Mercury, 'wiki.mainPageTitle'),
-			siteName: Em.getWithDefault(Mercury, 'wiki.siteName', 'Wikia')
+			siteName: Em.getWithDefault(Mercury, 'wiki.siteName', 'Wikia'),
+			editorPreview: Em.get(Mercury, 'article.preview')
 		});
 
 		// This event is for tracking mobile sessions between Mercury and WikiaMobile
