+++ front/templates/main/application.hbs
@@ -1,12 +1,22 @@
-{{smart-banner isVisible=view.smartBannerVisible}}
-{{site-head smartBannerVisible=view.smartBannerVisible sideNavCollapsed=view.sideNavCollapsed trackClick='trackClick'}}
-{{side-nav isCollapsed=view.sideNavCollapsed loadRandomArticle='loadRandomArticle'}}
-{{ad-slot name='MOBILE_TOP_LEADERBOARD' noAds=noAds}}
-<div class='page-wrapper' {{bind-attr lang=language.content}} {{bind-attr dir=language.contentDir}}>
-	<section class='wiki-container'>
-		<h2 class='wiki-title' {{action 'trackClick' 'wikiname'}}>{{#link-to 'article' mainPageTitle alt=siteName}}{{siteName}}{{/link-to}}</h2>
-		{{outlet}}
-	</section>
-</div>
-{{loading-spinner active=isLoading}}
-{{outlet 'lightbox'}}
+{{#if editorPreview}}
+	<div class="page-wrapper editor-preview" {{bind-attr lang=language.content}} {{bind-attr dir=language.contentDir}}>
+		<section class='wiki-container'>
+			{{outlet}}
+		</section>
+	</div>
+{{else}}
+	{{smart-banner isVisible=view.smartBannerVisible}}
+	{{site-head smartBannerVisible=view.smartBannerVisible sideNavCollapsed=view.sideNavCollapsed trackClick='trackClick'}}
+	{{side-nav isCollapsed=view.sideNavCollapsed loadRandomArticle='loadRandomArticle'}}
+	{{ad-slot name='MOBILE_TOP_LEADERBOARD' noAds=noAds}}
+	<div class='page-wrapper' {{bind-attr lang=language.content}} {{bind-attr dir=language.contentDir}}>
+		<section class='wiki-container'>
+			{{#unless editorPreview}}
+				<h2 class='wiki-title' {{action 'trackClick' 'wikiname'}}>{{#link-to 'article' mainPageTitle alt=siteName}}{{siteName}}{{/link-to}}</h2>
+			{{/unless}}
+			{{outlet}}
+		</section>
+	</div>
+	{{loading-spinner active=isLoading}}
+	{{outlet 'lightbox'}}
+{{/if}}
