+++ server/facets/editorPreview.ts
@@ -1,0 +1,50 @@
+/// <reference path="../../typings/hapi/hapi.d.ts" />
+
+import Article = require('../lib/Article');
+import Utils = require('../lib/Utils');
+import localSettings = require('../../config/localSettings');
+import verifyMWHash = require('./operations/verifyMWHash');
+import prepareArticleData = require('./operations/prepareArticleData');
+
+function editorPreview (request: Hapi.Request, reply: Hapi.Response): void {
+	var wikiDomain: string = Utils.getCachedWikiDomainName(localSettings, request.headers.host),
+		parserOutput: string = request.payload.parserOutput,
+		mwHash: string = request.payload.mwHash;
+
+	Article.getWikiVariables(wikiDomain, (error: any, wikiVariables: any) => {
+		var article: any = {},
+			result: any = {},
+			error: any;
+
+		if (verifyMWHash(parserOutput, mwHash)) {
+			article = JSON.parse(parserOutput);
+		} else {
+			error = true;
+		}
+
+		result = {
+			article: {
+				article: article,
+				adsContext: {},
+				details: {
+					id: 0,
+					title: '',
+					revision: {},
+					type: 'article'
+				},
+				preview: true
+			},
+			wiki: wikiVariables || {},
+			error: error
+		};
+
+		prepareArticleData(request, result);
+
+		// TODO: why is this needed for the images to load?
+		result.tracking = localSettings.tracking;
+
+		reply.view('application', result);
+	});
+}
+
+export = editorPreview;
