+++ server/facets/showArticle.ts
@@ -5,7 +5,7 @@ import Utils = require('../lib/Utils');
 import Tracking = require('../lib/Tracking');
 import Caching = require('../lib/Caching');
 import localSettings = require('../../config/localSettings');
-import crypto = require('crypto');
+import prepareArticleData = require('./operations/prepareArticleData');
 
 var cachingTimes = {
 	enabled: false,
@@ -14,30 +14,11 @@ var cachingTimes = {
 	browserTTL: Caching.Interval.default
 };
 
-function verifyMwHash (parserOutput: string, mwHash: string) {
-	var hmac = crypto.createHmac('sha1', localSettings.mwPreviewSalt),
-		computedHash = hmac.update(parserOutput).digest('hex');
-
-	return (computedHash === mwHash);
-}
-
 function showArticle (request: Hapi.Request, reply: Hapi.Response): void {
 	var path: string = request.path,
 		wikiDomain: string = Utils.getCachedWikiDomainName(localSettings, request.headers.host);
 
-	if (path === '/editor_preview/') {
-		if (!verifyMwHash(request.payload.parserOutput, request.payload.mwHash)) {
-			reply.redirect(localSettings.redirectUrlOnNoData);
-			return;
-		}
-
-		Article.getPreview({
-			wikiDomain: wikiDomain,
-			title: request.payload.title
-		}, JSON.parse(request.payload.parserOutput), (error: any, result: any = {}) => {
-			onArticleResponse(request, reply, error, result);
-		});
-	} else if (path === '/' || path === '/wiki/') {
+	if (path === '/' || path === '/wiki/') {
 		Article.getWikiVariables(wikiDomain, (error: any, wikiVariables: any) => {
 			if (error) {
 				// TODO check error.statusCode and react accordingly
@@ -64,51 +45,6 @@ function showArticle (request: Hapi.Request, reply: Hapi.Response): void {
 }
 
 /**
- * Prepares article data to be rendered
- * TODO: clean up this function
- *
- * @param {Hapi.Request} request
- * @param result
- */
-function beforeArticleRender (request: Hapi.Request, result: any): void {
-	var title: string,
-		articleDetails: any,
-		userDir = 'ltr';
-
-	if (result.article.details) {
-		articleDetails = result.article.details;
-		title = articleDetails.cleanTitle ? articleDetails.cleanTitle : articleDetails.title;
-	} else if (request.params.title) {
-		title = request.params.title.replace(/_/g, ' ');
-	} else {
-		title = result.wiki.mainPageTitle.replace(/_/g, ' ');
-	}
-
-	if (result.article.article) {
-		// we want to return the article content only once - as HTML and not JS variable
-		result.articleContent = result.article.article.content;
-		delete result.article.article.content;
-	}
-
-	if (result.wiki.language) {
-		userDir = result.wiki.language.userDir;
-		result.isRtl = (userDir === 'rtl');
-	}
-
-	result.displayTitle = title;
-	result.isMainPage = (title === result.wiki.mainPageTitle.replace(/_/g, ' '));
-	result.canonicalUrl = result.wiki.basePath + result.wiki.articlePath + title.replace(/ /g, '_');
-	result.themeColor = Utils.getVerticalColor(localSettings, result.wiki.vertical);
-	result.queryParams = Utils.parseQueryParams(request.query);
-
-	if (localSettings.optimizely.enabled && !result.queryParams.noexternals) {
-		result.optimizelyScript = localSettings.optimizely.scriptPath +
-			(localSettings.environment === Utils.Environment.Prod ?
-			localSettings.optimizely.account : localSettings.optimizely.devAccount) + '.js';
-	}
-}
-
-/**
  * Handles article response from API
  *
  * @param {Hapi.Request} request
@@ -131,7 +67,13 @@ function onArticleResponse (request: Hapi.Request, reply: any, error: any, resul
 			result.error = JSON.stringify(error);
 		}
 
-		beforeArticleRender(request, result);
+		prepareArticleData(request, result);
+
+		if (localSettings.optimizely.enabled && !result.queryParams.noexternals) {
+			result.optimizelyScript = localSettings.optimizely.scriptPath +
+			(localSettings.environment === Utils.Environment.Prod ?
+				localSettings.optimizely.account : localSettings.optimizely.devAccount) + '.js';
+		}
 
 		response = reply.view('application', result);
 		response.code(code);
